requirements.txt
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.18.1
